refactor: Split testing structs from lib.rs
This commit splits testing structs and functions from lib.rs
